14. Przenisienie klasy Node do klasy IntLinkedList w formie zagnieżdżonej prywatnej klasy i usunięcie zbędnego pliku pliku Node.java
